+++ src/main/java/com/rs/math/geometry/shape/Circle.java
@@ -1,9 +1,9 @@
 package com.rs.math.geometry.shape;
 
 public class Circle {
-    public final Point center;
-    public final float radius;
-    public Circle(Point center, float radius) {
+    public final Point  center;
+    public final double radius;
+    public Circle(Point center, double radius) {
         this.center = center;
         this.radius = radius;
     }
