+++ src/main/java/com/rs/math/geometry/value/Vector.java
@@ -3,14 +3,14 @@ package com.rs.math.geometry.value;
 import com.rs.math.geometry.shape.Point;
 
 public class Vector {
-    public final float x;
-    public final float y;
-    Vector(float x, float y, float length) {
-        float l = length / (float) Math.hypot(x, y);
+    public final double x;
+    public final double y;
+    Vector(double x, double y, double length) {
+        double l = length / Math.hypot(x, y);
         this.x = x * l;
         this.y = y * l;
     }
-    public Vector(float x, float y) {
+    public Vector(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -19,14 +19,14 @@ public class Vector {
         this.y = to.y - from.y;
     }
 
-    public float length() {
-        return (float) Math.hypot(x, y);
+    public double length() {
+        return Math.hypot(x, y);
     }
 
-    public static float cross(Vector a, Vector b) {
+    public static double cross(Vector a, Vector b) {
         return a.x * b.y - a.y * b.x;
     }
-    public static float dot(Vector a, Vector b) {
+    public static double dot(Vector a, Vector b) {
         return a.x * b.x + a.y * b.y;
     }
 }
